server-app/requirements.txt
Flask
Flask-SocketIO
eventlet
Pillow
pycryptodome
gunicorn
requests